+++ tests/models/charge_generation/test_radiation_induced_dark_current.py
@@ -7,10 +7,14 @@
 
 """Tests for radiation induced dark current model."""
 
+from typing import Union
+
 import pytest
 
 from pyxel.detectors import (
+    CCD,
     CMOS,
+    CCDGeometry,
     Characteristics,
     CMOSGeometry,
     Environment,
@@ -19,32 +23,50 @@ from pyxel.detectors import (
 from pyxel.models.charge_generation import radiation_induced_dark_current
 
 
-@pytest.fixture
-def cmos_5x5() -> CMOS:
+@pytest.fixture(params=["ccd", "cmos"])
+def detector_5x5(request: pytest.FixtureRequest) -> Union[CCD, CMOS]:
     """Create a valid CMOS detector."""
-    detector = CMOS(
-        geometry=CMOSGeometry(
-            row=5,
-            col=5,
-            total_thickness=40.0,
-            pixel_vert_size=10.0,
-            pixel_horz_size=10.0,
-        ),
-        environment=Environment(temperature=80),
-        characteristics=Characteristics(),
-    )
+    if request.param == "cmos":
+        detector = CMOS(
+            geometry=CMOSGeometry(
+                row=5,
+                col=5,
+                total_thickness=40.0,
+                pixel_vert_size=10.0,
+                pixel_horz_size=10.0,
+            ),
+            environment=Environment(temperature=80),
+            characteristics=Characteristics(),
+        )
+
+    elif request.param == "ccd":
+        detector = CCD(
+            geometry=CCDGeometry(
+                row=5,
+                col=5,
+                total_thickness=40.0,
+                pixel_vert_size=10.0,
+                pixel_horz_size=10.0,
+            ),
+            environment=Environment(temperature=80),
+            characteristics=Characteristics(),
+        )
+
+    else:
+        raise NotImplementedError
+
     detector._readout_properties = ReadoutProperties(times=[1.0])
     return detector
 
 
-def test_radiation_induced_dark_current_valid(cmos_5x5: CMOS):
+def test_radiation_induced_dark_current_valid(detector_5x5: Union[CCD, CMOS]):
     """Test model 'radiation_induced_dark_current' with valid inputs."""
-    detector = cmos_5x5
+    detector = detector_5x5
 
     # Call the function with appropriate arguments
     depletion_volume = 64  # µm3
     annealing_time = 0.1  # weeks
-    displacement_dose = 50  # TeV/g
+    displacement_dose = 50 * 10  # TeV/g
     seed = 42
 
     radiation_induced_dark_current(